backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings
python-dotenv>=1.0.0
pypdf>=4.0.0
sentence-transformers>=2.5.0
google-genai>=1.63.0
groq>=0.4.2
requests>=2.31.0
numpy>=1.24.0
supabase>=2.3.0
gotrue>=2.3.0
